refactor: component names and paths
