+++ core/app/[locale]/(default)/compare/_components/add-to-cart/fragment.ts
@@ -3,7 +3,6 @@ import { graphql } from '~/client/graphql';
 export const CartFragment = graphql(`
   fragment CartFragment on Product {
     entityId
-    name
     inventory {
       isInStock
     }
