+++ core/app/[locale]/(default)/compare/_components/add-to-cart/index.tsx
@@ -7,12 +7,12 @@ import { toast } from 'react-hot-toast';
 
 import { Link } from '~/components/link';
 
-import { addToCart } from '../_actions/add-to-cart';
+import { addToCart } from '../../_actions/add-to-cart';
 
-import { AddToCart } from './add-to-cart';
 import { CartFragment } from './fragment';
+import { SubmitButton } from './submit-button';
 
-export const AddToCartForm = ({ product }: { product: FragmentOf<typeof CartFragment> }) => {
+export const AddToCart = ({ product }: { product: FragmentOf<typeof CartFragment> }) => {
   const t = useTranslations('Compare');
 
   return (
@@ -53,7 +53,7 @@ export const AddToCartForm = ({ product }: { product: FragmentOf<typeof CartFrag
     >
       <input name="product_id" type="hidden" value={product.entityId} />
       <input name="quantity" type="hidden" value={1} />
-      <AddToCart disabled={!product.inventory.isInStock} productName={product.name} />
+      <SubmitButton disabled={!product.inventory.isInStock} />
     </form>
   );
 };
